+++ CHANGELOG.md
@@ -112,6 +112,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix issue where mask values were not being used correctly while painting collisions tiles [@Mico27](https://github.com/Mico27)
 - Fix issue where sprites with a canvas size of 8px wide would have first four pixels set as transparent
 - Fix issue where input text in dropdown search fields was unreadable in dark themes
+- Fix issue causing scroll bounds to be ignored for scenes with width of 20 or height of 18 tiles [@pau-tomas](https://github.com/pau-tomas)
 
 ## [4.1.3] - 2024-09-16